update components to move state to index.js
