+++ pages/components/SeasonSelector.js
@@ -1,15 +1,8 @@
 import React, { useState }from 'react';
 
-export default function SeasonSelector({onChange}) {
-    const years = ['Seasons', '2003', '2004', '2005', '2006']
-    const [chosenSeason, setChosenSeason] = useState('')
-    const handleChangeSeason = (event) => {
-      let season = event.target.value
-      setChosenSeason(season)
-      onChange(season)
-  }
+export default function SeasonSelector({years=[]}) {
     return (
-        <select name="" id="" onChange={handleChangeSeason}>
+        <select name="" id="">
             {years.map((year) => {
               return  <option key={year} value={year}>{year}</option>
             })}
