+++ pages/index.js
@@ -3,11 +3,41 @@ import Image from 'next/image'
 import LebronLakersImg from '../public/lebron.png'
 import { Inter } from '@next/font/google'
 import styles from '@/styles/Home.module.css'
-import React from 'react'
+import {useState, useEffect} from 'react'
 import PlayerInfo from './components/PlayerInfo'
+import SeasonSelector from './components/SeasonSelector'
 const inter = Inter({ subsets: ['latin'] })
 
 export default function Home() {
+  //Use State
+  const [player, setPlayer] = useState('')
+  const [loading, setLoading] = useState(true)
+  const [season, setSeason] = useState('2022')
+  const years = ['Seasons', '2003', '2004', '2005', '2006']
+
+
+  //Use Effect
+   useEffect(() => {
+    getPlayer(season)
+  }, [season])
+
+//Request functions
+  async function getPlayer(season){
+    const response = await fetch(`api/getLebron?season=${season}`)
+    const data = await response.json()
+    setPlayer(data)
+    setLoading(false)
+  }
+
+    const handleGetSeason = () => {
+    setSeason('')
+  }
+
+  const handleChangeSeason = (event) => {
+      let season = event.target.value
+      setSeason(season)
+  }
+
   return (
     <>
       <Head>
@@ -17,9 +47,12 @@ export default function Home() {
         <link rel="icon" href="/favicon.ico" />
       </Head>
       <main className={styles.playerSection}>
-          <div className={styles.player}>
-            <div><Image src={LebronLakersImg} width="100%" height="80vh" alt="Image"/></div>
-            <PlayerInfo />
+          <div className={styles.player} onChange={handleChangeSeason}>
+            <div>
+              <Image src={LebronLakersImg} width="100%" height="80vh" alt="Image"/>
+            </div>
+            <SeasonSelector years={years}/>
+            <PlayerInfo firstName={player.firstName} lastName={player.lastName} position={player.position} points={player.points} assists={player.assists} rebounds={player.rebounds} blocks={player.blocks} steals={player.steals} team={player.team}/>
           </div>
       </main>
     </>
